--- Assets/First Level/Camera Behavior.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class CameraBehavior : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-	
-	}
-	
-	// Update is called once per frame
-	void Update () {
-	
-	}
-}
