--- Assets/First Level/Camera Behavior.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 6e5af5725b485c9498949d73e5e52a53
-timeCreated: 1479136121
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
