--- Assets/First Level/Character.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 85a439a60d33d4c4b918ece73e578220
-timeCreated: 1477579823
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
