--- Assets/First Level/Enemy.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a5a65530c2631054f8d70d4495312d93
-timeCreated: 1476368359
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
