--- Assets/First Level/Floor Properties.physicsMaterial2D.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8e04521097636844ebddd0fe9fb129c6
-timeCreated: 1478096036
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
