--- Assets/First Level/Ground.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d6888ff6a54e76044a1f027cb7ed59fb
-timeCreated: 1476368435
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
