--- Assets/First Level/Player.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8da4d1afc1f3efd4bb99bd9a57b57079
-timeCreated: 1476368442
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
