--- Assets/First Level/Test.cs
@@ -1,19 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class Test : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-
-       
-
-    }
-	
-	// Update is called once per frame
-	void Update () {
-        //NewBehaviourScript asdf = new NewBehaviourScript();
-        //asdf.GetComponent<Rigidbody2D>
-
-    }
-}
