--- Assets/First Level/Test.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 88ec6e24df51e49438f789af16ac3fc5
-timeCreated: 1477408239
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
