--- Assets/First Level/input_Controller.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 85af75a63e0a4644dae6d6f10b26444e
-timeCreated: 1488383708
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
